# Conflicts:
#	bible/src/main/java/com/library/bible/cart/repository/ICartRepository.java
#	bible/src/main/java/com/library/bible/cart/service/ICartService.java

+++ bible/src/main/java/com/library/bible/book/service/IBookService.java
@@ -14,18 +14,18 @@ public interface IBookService {
 	int getBookCount(int categoryid);
 	
 	List<Map<String, Object>> getBookListMap();
-	Map<String, Object> getBookInfoMap(int bookid);
+	Map<String, Object> getBookInfoMap(long bookid);
 	
 	List<Book> getBookList();
-	Book getBookInfo(int bookid);
+	Book getBookInfo(long bookid);
 	
 	Book updateBook(Book book, MultipartFile file);
 	
 	void insertBook(Book book, MultipartFile file);
 	void insertBooks(List<Book> books);
 	
-	void deleteBook(int bookid);
-	int deleteBook(int bookid, String author);
+	void deleteBook(long bookid);
+	int deleteBook(long bookid, String author);
 	
 	List<Map<String, Object>> getBooksByCategory(String category);
 	List<Map<String, Object>> getAllAuthor();
