+++ bible/src/main/java/com/library/bible/cart/service/ICartService.java
@@ -6,20 +6,20 @@ import com.library.bible.cart.model.Cart;
 
 public interface ICartService {
 	//get cart info
-	List<Cart> getAllCarts(int memId);
-	Cart getCart(int cartId);
-	int isBookInCart(int memId, int bookId);
+	List<Cart> getAllCarts(long memId);
+	Cart getCart(long cartId);
+	int isBookInCart(long memId, long bookId);
 	
 	//update, delete, add
-	void updateCart(int cartId, int newCount);
-	void updateCartByBookId(int bookId, int memId, int bookCount);
-	void deleteCart(int cartId);
-	void addCart(int bookId, int memId, int bookCount);
+	void updateCart(long cartId, int newCount);
+	void updateCartByBookId(long bookId, long memId, int bookCount);
+	void deleteCart(long cartId);
+	void addCart(long bookId, long memId, int bookCount);
 	//구매 시 장바구니에서 삭제되는 deleteCarts
 	void deleteCarts(List<Integer> cartIds);
 	
 	//calculate selected prices
-	List<Integer> getSelectedCartPrices(List<Integer> cartId);
-	int getSelectedTotalPrice(List<Integer> cartId);
+	List<Long> getSelectedCartPrices(List<Long> cartId);
+	long getSelectedTotalPrice(List<Long> cartId);
 	List<Cart> getSelectedCarts(List<Integer> cartIds);
 }
